Rename PictureController and Add Ability to Remove PictureTakenListener
