+++ doc/Javadoc/com/example/atheneum/controllers/PictureController.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.atheneum.controllers.PictureController&lt;T&gt;</li>
+<li>com.example.atheneum.utils.CameraHandler&lt;T&gt;</li>
 </ul>
 </li>
 </ul>
@@ -184,7 +184,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i4" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../com/example/atheneum/controllers/PictureController.html#setPictureTakenListener-com.example.atheneum.controllers.PictureController.OnPictureTakenListener-">setPictureTakenListener</a></span>(<a href="../../../../com/example/atheneum/controllers/PictureController.OnPictureTakenListener.html" title="interface in com.example.atheneum.controllers">PictureController.OnPictureTakenListener</a>&nbsp;pictureTakenListener)</code>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../com/example/atheneum/controllers/PictureController.html#setPictureTakenListener-com.example.atheneum.utils.CameraHandler.OnPictureTakenListener-">setPictureTakenListener</a></span>(<a href="../../../../com/example/atheneum/controllers/PictureController.OnPictureTakenListener.html" title="interface in com.example.atheneum.controllers">PictureController.OnPictureTakenListener</a>&nbsp;pictureTakenListener)</code>
 <div class="block">Sets the interface of callback(s) to handle events from controller</div>
 </td>
 </tr>
@@ -240,7 +240,7 @@ extends java.lang.Object</pre>
 </dl>
 </li>
 </ul>
-<a name="setPictureTakenListener-com.example.atheneum.controllers.PictureController.OnPictureTakenListener-">
+<a name="setPictureTakenListener-com.example.atheneum.utils.CameraHandler.OnPictureTakenListener-">
 <!--   -->
 </a>
 <ul class="blockList">
