+++ doc/Javadoc/index-files/index-16.html
@@ -188,7 +188,7 @@
 </dd>
 <dt><span class="memberNameLink"><a href="../com/example/atheneum/models/User.html#setPhotos-java.util.ArrayList-">setPhotos(ArrayList&lt;String&gt;)</a></span> - Method in class com.example.atheneum.models.<a href="../com/example/atheneum/models/User.html" title="class in com.example.atheneum.models">User</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../com/example/atheneum/controllers/PictureController.html#setPictureTakenListener-com.example.atheneum.controllers.PictureController.OnPictureTakenListener-">setPictureTakenListener(PictureController.OnPictureTakenListener)</a></span> - Method in class com.example.atheneum.controllers.<a href="../com/example/atheneum/controllers/PictureController.html" title="class in com.example.atheneum.controllers">PictureController</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/atheneum/controllers/PictureController.html#setPictureTakenListener-com.example.atheneum.utils.CameraHandler.OnPictureTakenListener-">setPictureTakenListener(PictureController.OnPictureTakenListener)</a></span> - Method in class com.example.atheneum.controllers.<a href="../com/example/atheneum/controllers/PictureController.html" title="class in com.example.atheneum.controllers">PictureController</a></dt>
 <dd>
 <div class="block">Sets the interface of callback(s) to handle events from controller</div>
 </dd>
